viewer: Drop rlib from cdylib lib to avoid Windows test linker race
The workspace sets `[profile.dev] panic = "abort"` while `[profile.test]`
keeps the default `panic = "unwind"`. With `crate-type = ["lib", "cdylib"]`
the bin's compilation auto-links the package's rlib, so cargo has to build
the lib for the bin's dev profile as well as for the test runner's profile.
The two builds emit the same hashless cdylib filename (`libslint_viewer.dll`),
and on Windows the parallel link.exe processes race on that file and fail
with LNK1104.

Drop `lib` from `crate-type` so cargo does not auto-link the package's
library into the bin. The cdylib is still produced for Android's
NativeActivity. `poll_ready` and `mod remote` move into the bin since
they're no longer reachable through `slint_viewer::`.
