+++ tools/viewer/lib.rs
@@ -1,22 +1,14 @@
 // Copyright © SixtyFPS GmbH <info@slint.dev>
 // SPDX-License-Identifier: GPL-3.0-only OR LicenseRef-Slint-Royalty-free-2.0 OR LicenseRef-Slint-Software-3.0
 
-#[cfg(feature = "remote")]
-pub mod remote;
-
-/// Poll a future that is expected to resolve immediately (e.g. the interpreter's
-/// `build_from_path` when no async file loader is installed).
-pub fn poll_ready<F: std::future::Future>(future: F) -> F::Output {
-    let mut future = core::pin::pin!(future);
-    let mut cx = std::task::Context::from_waker(std::task::Waker::noop());
-    match std::future::Future::poll(future.as_mut(), &mut cx) {
-        std::task::Poll::Ready(result) => result,
-        std::task::Poll::Pending => unreachable!("Compiler returned Pending"),
-    }
-}
+// This file exists only to expose `android_main` from the cdylib for Android's
+// NativeActivity. The viewer's command-line implementation lives in main.rs.
 
 #[cfg(all(target_os = "android", not(feature = "remote")))]
 compile_error!("The `remote` feature is required when building for Android");
+
+#[cfg(all(target_os = "android", feature = "remote"))]
+mod remote;
 
 #[cfg(all(target_os = "android", feature = "remote"))]
 #[unsafe(no_mangle)]
