+++ tools/viewer/screenshot.rs
@@ -10,7 +10,7 @@ use std::cell::Cell;
 use std::path::Path;
 use std::rc::Rc;
 
-use crate::{Cli, Error, Result, extract_component, init_compiler, setup_instance};
+use crate::{Cli, Error, Result, extract_component, init_compiler, poll_ready, setup_instance};
 
 /// Build the best headless renderer compiled into the viewer. Skia's software
 /// rasterizer is preferred when available; otherwise we fall back to Slint's
@@ -52,7 +52,7 @@ pub fn take_screenshot(args: &Cli) -> Result<()> {
     }
 
     let compiler = init_compiler(args);
-    let result = slint_viewer::poll_ready(compiler.build_from_path(args.path()));
+    let result = poll_ready(compiler.build_from_path(args.path()));
     result.print_diagnostics();
     if result.has_errors() {
         std::process::exit(-1);
